requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pandas==2.2.0 